Module darkfi_money_contract::client

Client API for interaction with this smart contract This module implements the client-side API for this contract’s interaction. What we basically do here is implement an API that creates the necessary structures and is able to export them to create a DarkFi transaction object that can be broadcasted to the network when we want to make a payment with some coins in our wallet.

Note that this API does not involve any wallet interaction, but only takes the necessary objects provided by the caller. This is intentional, so we are able to abstract away any wallet interfaces to client implementations.

Modules

genesis_mint_v1

Money::GenesisMintV1 API

stake_v1

Money::StakeV1 API This API is crufty. Please rework it into something nice to read and nice to use.

swap_v1

Money::OtcSwapV1 API This API is crufty. Please rework it into something nice to read and nice to use.

token_freeze_v1

Money::TokenFreezeV1 API

token_mint_v1

Money::TokenMintV1 API

transfer_v1

Money::TransferV1 API This API is crufty. Please rework it into something nice to read and nice to use.

unstake_v1

Money::UnstakeV1 API This API is crufty. Please rework it into something nice to read and nice to use.

Structs

ConsensusNote

ConsensusNote holds the inner attributes of a Coin.

ConsensusOwnCoin

ConsensusOwnCoin is a representation of Coin with its respective metadata.

MoneyNote

MoneyNote holds the inner attributes of a Coin.

OwnCoin

OwnCoin is a representation of Coin with its respective metadata.

Constants

MONEY_ALIASES_COL_ALIAS

MONEY_ALIASES_COL_TOKEN_ID

MONEY_ALIASES_TABLE

MONEY_COINS_COL_COIN

MONEY_COINS_COL_IS_SPENT

MONEY_COINS_COL_LEAF_POSITION

MONEY_COINS_COL_MEMO

MONEY_COINS_COL_NULLIFIER

MONEY_COINS_COL_SECRET

MONEY_COINS_COL_SERIAL

MONEY_COINS_COL_SPEND_HOOK

MONEY_COINS_COL_TOKEN_BLIND

MONEY_COINS_COL_TOKEN_ID

MONEY_COINS_COL_USER_DATA

MONEY_COINS_COL_VALUE

MONEY_COINS_COL_VALUE_BLIND

MONEY_COINS_TABLE

MONEY_INFO_COL_LAST_SCANNED_SLOT

MONEY_INFO_TABLE

MONEY_KEYS_COL_IS_DEFAULT

MONEY_KEYS_COL_KEY_ID

MONEY_KEYS_COL_PUBLIC

MONEY_KEYS_COL_SECRET

MONEY_KEYS_TABLE

MONEY_TOKENS_COL_IS_FROZEN

MONEY_TOKENS_COL_MINT_AUTHORITY

MONEY_TOKENS_COL_TOKEN_ID

MONEY_TOKENS_TABLE

MONEY_TREE_COL_TREE

MONEY_TREE_TABLE