Module darkfi_money_contract::client::transfer_v1

Money::TransferV1 API This API is crufty. Please rework it into something nice to read and nice to use.

Structs

TransferBurnRevealed

TransferCallBuilder

Struct holding necessary information to build a Money::TransferV1 contract call.

TransferCallBuilder2

Struct holding necessary information to build a Money::TransferV1 contract call.

TransferCallClearInput

TransferCallDebris

Output metadata claimed from building a Money::Transfer call

TransferCallInput

TransferCallOutput

TransferCallSecrets 🔒

TransferMintRevealed

Functions

compute_remainder_blind 🔒

create_transfer_burn_proof

create_transfer_mint_proof

select_coins

Select coins from coins of at least min_value in total. Different strategies can be used. This function uses the dumb strategy of selecting coins until we reach min_value.