Crate darkfi_money_contract

Smart contract implementing money transfers, atomic swaps, token minting and freezing, and staking/unstaking of consensus tokens.

Modules

client

Client API for interaction with this smart contract This module implements the client-side API for this contract’s interaction. What we basically do here is implement an API that creates the necessary structures and is able to export them to create a DarkFi transaction object that can be broadcasted to the network when we want to make a payment with some coins in our wallet.

error

Internal contract errors

model

Call parameters definitions

Enums

MoneyFunction

Functions available in the contract

Constants

EMPTY_COINS_TREE_ROOT

Precalculated root hash for a tree containing only a single Fp::ZERO coin. Used to save gas.

MONEY_CONTRACT_COINS_TREE

MONEY_CONTRACT_COIN_MERKLE_TREE

MONEY_CONTRACT_COIN_ROOTS_TREE

MONEY_CONTRACT_DB_VERSION

MONEY_CONTRACT_FEES_TREE

MONEY_CONTRACT_INFO_TREE

MONEY_CONTRACT_LATEST_COIN_ROOT

MONEY_CONTRACT_LATEST_NULLIFIER_ROOT

MONEY_CONTRACT_NULLIFIERS_TREE

MONEY_CONTRACT_NULLIFIER_ROOTS_TREE

MONEY_CONTRACT_TOKEN_FREEZE_TREE

MONEY_CONTRACT_ZKAS_AUTH_TOKEN_MINT_NS_V1

zkas token auth mint circuit namespace

MONEY_CONTRACT_ZKAS_BURN_NS_V1

zkas burn circuit namespace

MONEY_CONTRACT_ZKAS_FEE_NS_V1

zkas fee circuit namespace

MONEY_CONTRACT_ZKAS_MINT_NS_V1

zkas mint circuit namespace

MONEY_CONTRACT_ZKAS_TOKEN_MINT_NS_V1

zkas token mint circuit namespace