Module darkfi_money_contract::client

Client API for interaction with this smart contract This module implements the client-side API for this contract’s interaction. What we basically do here is implement an API that creates the necessary structures and is able to export them to create a DarkFi transaction object that can be broadcasted to the network when we want to make a payment with some coins in our wallet.

Note that this API does not involve any wallet interaction, but only takes the necessary objects provided by the caller. This is intentional, so we are able to abstract away any wallet interfaces to client implementations.

Modules

auth_token_freeze_v1

Money::AuthTokenFreezeV1 API

auth_token_mint_v1

Money::AuthTokenMintV1 API

fee_v1

Money::FeeV1 API

genesis_mint_v1

Money::GenesisMintV1 API

pow_reward_v1

Money::PoWRewardV1 API

swap_v1

Money::OtcSwapV1 API This API is crufty. Please rework it into something nice to read and nice to use.

token_mint_v1

Money::TokenMintV1 API

transfer_v1

Money::TransferV1 API

Structs

MoneyNote

MoneyNote holds the inner attributes of a Coin.

OwnCoin

OwnCoin is a representation of Coin with its respective metadata.

Functions

compute_remainder_blind