Module darkfi_money_contract::client

Transaction building API for clients interacting with this contract. This module implements the client-side of this contract’s interaction. What we basically do here is implement an API that creates the necessary structures and is able to export them to create a DarkFi Transaction object that can be broadcasted to the network when we want to make a payment with some coins in our wallet. Note that this API doesn’t involve any wallet interaction, but only takes the necessary objects provided by the caller. This is so we can abstract away the wallet interface to client implementations.

Structs

Coin

The Coin is represented as a base field element.

EncryptedNote

The EncryptedNote represents a structure holding the ciphertext (which is an encryption of the Note object, and the ephemeral PublicKey created at the time when the encryption was done

Note

The Note holds the inner attributes of a Coin

OwnCoin

The OwnCoin is a representation of Coin with its respective metadata.

StakeLeadMintRevealed 🔒

TransactionBuilderClearInputInfo 🔒

TransactionBuilderInputInfo 🔒

TransactionBuilderOutputInfo 🔒

TransferBurnRevealed

TransferMintRevealed

UnstakeLeadBurnRevealed 🔒

Constants

AEAD_TAG_SIZE

Byte length of the AEAD tag of the chacha20 cipher used for note encryption

MONEY_COINS_COL_COIN

MONEY_COINS_COL_COIN_BLIND

MONEY_COINS_COL_IS_SPENT

MONEY_COINS_COL_LEAF_POSITION

MONEY_COINS_COL_MEMO

MONEY_COINS_COL_NULLIFIER

MONEY_COINS_COL_SECRET

MONEY_COINS_COL_SERIAL

MONEY_COINS_COL_SPEND_HOOK

MONEY_COINS_COL_TOKEN_BLIND

MONEY_COINS_COL_TOKEN_ID

MONEY_COINS_COL_USER_DATA

MONEY_COINS_COL_VALUE

MONEY_COINS_COL_VALUE_BLIND

MONEY_COINS_TABLE

MONEY_INFO_COL_LAST_SCANNED_SLOT

MONEY_INFO_TABLE

MONEY_KEYS_COL_IS_DEFAULT

MONEY_KEYS_COL_KEY_ID

MONEY_KEYS_COL_PUBLIC

MONEY_KEYS_COL_SECRET

MONEY_KEYS_TABLE

MONEY_TREE_COL_TREE

MONEY_TREE_TABLE

Functions

build_half_swap_tx

Build half of the money contract OTC swap transaction parameters with the given data:

build_stake_tx

build_transfer_tx

Build money contract transfer transaction parameters with the given data:

build_unstake_tx

compute_remainder_blind 🔒

create_stake_mint_proof 🔒

create_transfer_burn_proof

create_transfer_mint_proof

create_unstake_burn_proof 🔒